Refactor: libcrmcommon: Move some functions to patchset_display.c
Move the functions that are used for output/logging, in preparation for
converting some of the patchset output functions to message formatter
functions.

No code changes.

In other libraries so far, we've put message functions in a file called
"<lib_prefix>_output.c". However, in libcrmcommon, using "output" in the
filename is confusing, because we have several output*.c functions that
implement the pcmk__output_t API. We've already named an XML output file
"xml_display.c" (to distinguish it from output_xml.c), so we'll use
"patchset_display.c" for consistency.

Furthermore, having a single message function file per library has
become unwieldy in the past, so it's better to split them up into
logical groups like patchset functions. This is why we'll use
"patchset_display.c" instead of a monolithic file called (for example)
"common_output.c."
